Add an SPDT switch
This can be used both as a switch or a jumper, if using a jumper we can
hook a low value resistor to measure the load on the power supply.

The routing for the power line goes straight through this switch so even
a programming power input can be switched on/off.